Initial exploration of the NSUB
This NIP exploration is an attempt to solve a real world problem with key management and account security and loss prevention. I'll need devs to pick this apart and help it align to protocol goals and abilities. End goal is to allow secure user and entity adoption of the protocol without any steering toward centralization.